requirements.txt
transformers==4.40.1
datasets==2.18.0
scikit-learn==1.3.2
pandas==2.2.2
torch>=2.1
matplotlib==3.8.4
tqdm>=4.66.0
numpy==1.26.4
google-genai>=1.0.0
lime
shap